Moved to openapi/types.ts from openapi.ts
